refactor: extract optional and required claims into separate files
